Get user information only once upon login.
